huaweicloud/src/requirements.txt
cloudshell-shell-core>=4.0.0,<4.1.0
cloudshell-cp-core>=1.0.0,<1.1.0
cloudshell-automation-api>=9.0.0,<9.1.0
openstacksdk==0.17.1
